move print_layer_structure and print_interface_states to print_structure module
